requirements.txt
litellm
loguru
openai
python-dotenv
httpx[socks]